Add AbstractBrowserItem class for browser item hierarchy
Introduces the `AbstractBrowserItem` class in the
`electrifier.Controls.Contracts` namespace as an abstract
base for browser items of type `T`. This class includes
properties for child items, folder status, and root item
status, along with a constructor and a customizable
`ToString` method. Updates to `BrowserItemFactory`
reflect the removal of the previous abstract class,
centralizing browser item functionality.
